requirements.txt
Flask
tensorflow
numpy
pandas
matplotlib
Pillow
sklearn
gunicorn
shutil
werkzeug
opencv-python
